requirements.txt
#para instalar requerimientos
#pip install -r requirements.txt

astroid==2.2.5
colorama==0.4.1
Django==2.2
django-filter==2.1.0
django-mysql==3.0.0.post1
djangorestframework==3.9.4
isort==4.3.17
lazy-object-proxy==1.3.1
Markdown==3.1.1
mccabe==0.6.1
Pillow==6.0.0
pylint==2.3.1
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
typed-ast==1.3.4
virtualenv==16.5.0
wrapt==1.11.1

